+++ sql/BUILD
@@ -21,7 +21,8 @@ ts_test_suite(
     name = "tests",
     srcs = glob(["**/*.spec.ts"]),
     data = [
-        "//test_credentials",
+        "//test_credentials:bigquery.json",
+        "//test_credentials:snowflake.json",
         "@npm//source-map-support",
     ],
     templated_args = ["--node_options=--require=source-map-support/register"],
