+++ test_credentials/BUILD
@@ -1,13 +1,10 @@
-package(default_visibility = [
-    "//visibility:public",
-])
-
-package(default_visibility = ["//tests:__subpackages__"])
-
 load("//tools/gcloud:secrets.bzl", "gcloud_secret")
+
+package(default_visibility = ["//visibility:public"])
 
 gcloud_secret(
     name = "bigquery.json",
+    testonly = 1,
     ciphertext_file = ":bigquery.json.enc",
     key = "dataform-co-build",
     keyring = "dataform-co-build",
@@ -15,6 +12,7 @@ gcloud_secret(
 
 gcloud_secret(
     name = "redshift.json",
+    testonly = 1,
     ciphertext_file = ":redshift.json.enc",
     key = "dataform-co-build",
     keyring = "dataform-co-build",
@@ -22,6 +20,7 @@ gcloud_secret(
 
 gcloud_secret(
     name = "snowflake.json",
+    testonly = 1,
     ciphertext_file = ":snowflake.json.enc",
     key = "dataform-co-build",
     keyring = "dataform-co-build",
@@ -29,6 +28,7 @@ gcloud_secret(
 
 gcloud_secret(
     name = "sqldatawarehouse.json",
+    testonly = 1,
     ciphertext_file = ":sqldatawarehouse.json.enc",
     key = "dataform-co-build",
     keyring = "dataform-co-build",
